refactor(server): reorganize module structure and consolidate utilities
- Move server utilities from `server::utils` to `server::core::utils` for better organization
- Relocate metadata handling from `runtime::metadata` to `server::rendering::metadata`
- Update all import paths across RSC actions, rendering, runtime, and ops modules
- Reorganize Node.js runtime initialization with dynamic version and process properties
- Consolidate v8_utils imports in deno_runtime to reduce redundant use statements
- Remove metadata module from runtime and import from server rendering instead
- Update Node.js process object to use Deno runtime properties when available (execPath, pid, ppid)
- Replace hardcoded Node version strings with placeholders for dynamic substitution
- Improves code organization by centralizing server-related utilities and cross-module concerns
