+++ crates/rari/src/rsc/actions/mod.rs
@@ -165,7 +165,7 @@ fn check_origin(headers: &HeaderMap, allowed_origins: &[String]) -> Result<(), S
     }
 
     if let Some(origin) = headers.get("origin").and_then(|v| v.to_str().ok()) {
-        if !crate::server::utils::http::is_origin_allowed(origin, allowed_origins) {
+        if !crate::server::core::utils::http::is_origin_allowed(origin, allowed_origins) {
             error!("Invalid origin: {}", origin);
             return Err(StatusCode::FORBIDDEN);
         }
@@ -181,7 +181,8 @@ fn check_origin(headers: &HeaderMap, allowed_origins: &[String]) -> Result<(), S
                 } else {
                     format!("{}://{}:{}", scheme, host, port)
                 };
-            if crate::server::utils::http::is_origin_allowed(&referer_origin, allowed_origins) {
+            if crate::server::core::utils::http::is_origin_allowed(&referer_origin, allowed_origins)
+            {
                 return Ok(());
             }
             error!("Invalid referer origin: {}", referer_origin);
@@ -505,7 +506,7 @@ pub async fn handle_form_action(
                             "{}://{}:{}",
                             referer_tuple.0, referer_tuple.1, referer_tuple.2
                         );
-                        let allowed = crate::server::utils::http::is_origin_allowed(
+                        let allowed = crate::server::core::utils::http::is_origin_allowed(
                             &referer_origin,
                             &allowed_origins,
                         );
