+++ crates/rari/src/rsc/rendering/layout/utils.rs
@@ -1,6 +1,6 @@
+use crate::server::core::utils::component::{readable_component_id, short_hash};
 use crate::server::routing::app_router::AppRouteMatch;
 use crate::server::routing::types::ParamValue;
-use crate::server::utils::component::{readable_component_id, short_hash};
 use cow_utils::CowUtils;
 use rari_error::RariError;
 use rustc_hash::FxHashMap;
