+++ crates/rari/src/runtime/utils/transpile.rs
@@ -1,7 +1,9 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
+use std::borrow::Cow;
 use std::path::Path;
 
+use cow_utils::CowUtils;
 use deno_ast::MediaType;
 use deno_ast::ParseParams;
 use deno_ast::SourceMapOption;
@@ -13,11 +15,41 @@ use deno_error::JsErrorBox;
 deno_error::js_error_wrapper!(deno_ast::ParseDiagnostic, JsParseDiagnostic, "Error");
 deno_error::js_error_wrapper!(deno_ast::TranspileError, JsTranspileError, "Error");
 
+/// NOTE: In deno_node >= 0.190.0, this is exported as `deno_node::NODE_VERSION`.
+/// When we upgrade, we should use that constant directly to stay aligned.
+const NODE_VERSION: &str = "26.3.0";
+
+fn maybe_substitute_version_placeholders(name: &str, source: ModuleCodeString) -> ModuleCodeString {
+    const NODE_VERSION_TOKEN: &str = "__NODE_VERSION__";
+    const V8_VERSION_TOKEN: &str = "__V8_VERSION__";
+
+    if !name.contains("ext/node/init_node.js") && !name.contains("ext:node/init_node.js") {
+        return source;
+    }
+
+    let source_str = source.as_str();
+
+    if !source_str.contains(NODE_VERSION_TOKEN) && !source_str.contains(V8_VERSION_TOKEN) {
+        return source;
+    }
+
+    let result: Cow<str> = source_str.cow_replace(NODE_VERSION_TOKEN, NODE_VERSION);
+
+    if result.contains(V8_VERSION_TOKEN) {
+        let final_result: Cow<str> =
+            result.cow_replace(V8_VERSION_TOKEN, deno_core::v8::VERSION_STRING);
+        return final_result.into_owned().into();
+    }
+
+    result.into_owned().into()
+}
+
 pub fn maybe_transpile_source(
     name: ModuleName,
     source: ModuleCodeString,
 ) -> Result<(ModuleCodeString, Option<SourceMapData>), JsErrorBox> {
-    // Always transpile `node:` built-in modules, since they might be TypeScript.
+    let name_string = name.to_string();
+    let source = maybe_substitute_version_placeholders(&name_string, source);
     let media_type = if name.starts_with("node:") {
         MediaType::TypeScript
     } else {
