+++ crates/rari/src/server/core/utils/client.rs
