+++ crates/rari/src/server/core/utils/component.rs
