+++ crates/rari/src/server/core/utils/http.rs
