+++ crates/rari/src/server/core/utils/path_validation.rs
