+++ crates/rari/src/server/handlers/api.rs
@@ -1,6 +1,7 @@
 use crate::server::ServerState;
+use crate::server::core::utils::http::{add_api_cors_headers, add_api_security_headers};
 use crate::server::handlers::r#static::cors_preflight_response;
-use crate::server::utils::http::{add_api_cors_headers, add_api_security_headers};
+use crate::server::routing::api_error::{ApiRouteError, create_generic_error_response};
 use axum::{
     body::Body,
     extract::State,
@@ -65,8 +66,6 @@ pub async fn handle_api_route(
     State(state): State<ServerState>,
     req: axum::http::Request<Body>,
 ) -> Result<axum::http::Response<Body>, StatusCode> {
-    use crate::server::routing::api_error::{ApiRouteError, create_generic_error_response};
-
     let path = req.uri().path().to_string();
     let method = req.method().to_string();
     let is_development = state.config.is_development();
