+++ crates/rari/src/server/middleware/request_context.rs
@@ -1,4 +1,4 @@
-use crate::server::utils::client::get_http_client;
+use crate::server::core::utils::client::get_http_client;
 use axum::http::HeaderMap;
 use bytes::Bytes;
 use dashmap::DashMap;
