+++ crates/rari/src/server/rendering/metadata.rs
