+++ crates/rari/src/server/rendering/mod.rs
@@ -1,3 +1,4 @@
+pub mod metadata;
 pub mod metadata_injection;
 pub mod streaming_response;
 pub mod utils;
